[mem] 64bit: fix incosistent cast - unsigned long was used for storing a mem object size, but an int was used for the cast fetching it (which is wrong on 64bit)

+++ mem/tmalloc.c
@@ -99,7 +99,7 @@ static void obj_registry_stats(void)
 
     for (rec = obj_registry; rec; rec = rec->next)
     {
-        int sz = *(((int *)rec->p) - 1);
+        int sz = *(((unsigned long *)rec->p) - 1);
 
         tp_out(("%p size %d type %s\n", rec->p, sz, rec->type));
     }
